back_office/requirements.txt
Django==3.2.3
Pillow==9.4.0
requests==2.25.1
celery==5.2.7
django-cors-headers==3.13.0